[ipSDK] BuildHelper 수정  - 설정 삭제 및 Job 복사 기능 추가.
git-svn-id: https://192.168.6.225:8443/svn/CAD/Tag20120904/Utils/BuildHelper@42678 d8ace140-e149-8147-8d3d-8bcd6e031801

+++ BuildHelper/BuildHelper.cpp
@@ -266,7 +266,7 @@ bool ConsoleMode()
 	}
 	else if (params[0].CompareNoCase(L"run") == 0)
 	{
-		CBuildHelperApp::CmdRun(params.size(), &params[0]);
+		CBuildHelperApp::CmdRun((int)params.size(), &params[0]);
 	}
 	else if (params[0].CompareNoCase(L"delete") == 0)
 	{
